+++ addons/yat/src/scenes/overlay/components/terminal/components/terminal_context/quick_commands/QuickCommands.cs
@@ -1,0 +1,10 @@
+using Godot;
+using Godot.Collections;
+
+namespace YAT.Scenes.Overlay.Components.Terminal
+{
+	public partial class QuickCommands : Resource
+	{
+		[Export] public Dictionary<string, string> Commands { get; set; } = new();
+	}
+}